requirements.txt
# Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# RAG and AI
langchain==0.1.20
langchain-openai==0.1.7
langchain-community==0.0.38
langchain-core==0.1.52
chromadb==0.4.22
sentence-transformers==2.5.1
openai==1.40.0

# Document Processing
pypdf==3.17.4
python-docx==1.1.0
unstructured==0.11.6

# Frontend
streamlit==1.29.0
streamlit-chat==0.1.1

# Utilities
python-dotenv==1.0.0
numpy==1.26.2
pandas==2.1.4

